Fix code climate issues
